backend/requirements.txt
fastapi
uvicorn
python-jose
email-validator
python-dotenv
sqlalchemy
alembic
psycopg[binary]
passlib[bcrypt]
